Restored latest cap file endpoint and removed latest alert one
