+++ deploy_tools/nginx.template.conf
@@ -3,7 +3,7 @@ server {
 	listen [::]:80 default_server;
 
     # Path to the built react app
-	root /var/www/<site_name>;
+	root /var/www/rss;
 	index index.html;
 
 	server_name _;
