+++ src/rss/api/alerts/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+api = Blueprint("api", __name__)
+
+from . import alerts
