+++ src/rss/api/errors/errors.py
